# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resume Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/resumes/extract/pdf": {
      "post": {
        "summary": "Pdf",
        "operationId": "pdf_api_resumes_extract_pdf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pdf_api_resumes_extract_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractPageDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/extract/image": {
      "post": {
        "summary": "Image",
        "operationId": "image_api_resumes_extract_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_image_api_resumes_extract_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractPageDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/extract/docx": {
      "post": {
        "summary": "Docx",
        "operationId": "docx_api_resumes_extract_docx_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_docx_api_resumes_extract_docx_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractFlowDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/extract/txt": {
      "post": {
        "summary": "Txt",
        "operationId": "txt_api_resumes_extract_txt_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_txt_api_resumes_extract_txt_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractFlowDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resumes/extract/text": {
      "post": {
        "summary": "Text",
        "operationId": "text_api_resumes_extract_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractFlowDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_docx_api_resumes_extract_docx_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_docx_api_resumes_extract_docx_post"
      },
      "Body_image_api_resumes_extract_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_image_api_resumes_extract_image_post"
      },
      "Body_pdf_api_resumes_extract_pdf_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_pdf_api_resumes_extract_pdf_post"
      },
      "Body_txt_api_resumes_extract_txt_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_txt_api_resumes_extract_txt_post"
      },
...
      "DirectTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DirectTextRequest"
      },
      "DocumentPage": {
        "properties": {
          "pageNumber": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Pagenumber"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/PageBlock"
            },
            "type": "array",
            "minItems": 1,
            "title": "Blocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pageNumber",
          "blocks"
        ],
        "title": "DocumentPage"
      },
...
      "ExtractFlowDocumentResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/FlowDocument"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleErrorDTO"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document",
          "errors"
        ],
        "title": "ExtractFlowDocumentResponse"
      },
      "ExtractPageDocumentResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/PageDocument"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleErrorDTO"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document",
          "errors"
        ],
        "title": "ExtractPageDocumentResponse"
      },
      "ExtractionIssue": {
        "properties": {
          "stage": {
            "type": "string",
            "enum": [
              "assessment",
              "recovery"
            ],
            "title": "Stage"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "pageNumber": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pagenumber"
          },
          "lineIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lineids"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NormalizedBBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "recovered": {
            "type": "boolean",
            "title": "Recovered",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stage",
          "code",
          "message"
        ],
        "title": "ExtractionIssue"
      },
      "ExtractionReport": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "recovered",
              "needs_review"
            ],
            "title": "Status"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ExtractionIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ExtractionReport"
      },
...
      "FlowBlock": {
        "properties": {
          "blockId": {
            "type": "string",
            "pattern": "^f-b[1-9]\\d*$",
            "title": "Blockid"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/FlowLine"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "blockId",
          "lines"
        ],
        "title": "FlowBlock"
      },
      "FlowDocument": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/FlowBlock"
            },
            "type": "array",
            "minItems": 1,
            "title": "Blocks"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "blocks"
        ],
        "title": "FlowDocument"
      },
      "FlowLine": {
        "properties": {
          "lineId": {
            "type": "string",
            "pattern": "^f-l[1-9]\\d*$",
            "title": "Lineid"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "startOffset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Startoffset"
          },
          "endOffset": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Endoffset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lineId",
          "text",
          "startOffset",
          "endOffset"
        ],
        "title": "FlowLine"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModuleErrorDTO": {
        "properties": {
          "moduleKey": {
            "type": "string",
            "title": "Modulekey"
          },
          "inputSourceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputsourceid"
          },
          "errorCode": {
            "type": "string",
            "title": "Errorcode"
          },
          "userMessage": {
            "type": "string",
            "title": "Usermessage"
          },
          "canRetry": {
            "type": "boolean",
            "title": "Canretry"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "moduleKey",
          "inputSourceId",
          "errorCode",
          "userMessage",
          "canRetry"
        ],
        "title": "ModuleErrorDTO"
      },
...
      "NormalizedBBox": {
        "properties": {
          "x": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Y"
          },
          "width": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "NormalizedBBox"
      },
      "PageBlock": {
        "properties": {
          "blockId": {
            "type": "string",
            "pattern": "^p[1-9]\\d*-b[1-9]\\d*$",
            "title": "Blockid"
          },
          "bbox": {
            "$ref": "#/components/schemas/NormalizedBBox"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PageLine"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "blockId",
          "bbox",
          "lines"
        ],
        "title": "PageBlock"
      },
      "PageDocument": {
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/DocumentPage"
            },
            "type": "array",
            "minItems": 1,
            "title": "Pages"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pages"
        ],
        "title": "PageDocument"
      },
      "PageLine": {
        "properties": {
          "lineId": {
            "type": "string",
            "pattern": "^p[1-9]\\d*-l[1-9]\\d*$",
            "title": "Lineid"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "bbox": {
            "$ref": "#/components/schemas/NormalizedBBox"
          },
          "textSource": {
            "type": "string",
            "enum": [
              "embedded",
              "ocr"
            ],
            "title": "Textsource"
          },
          "uncertainWords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Uncertainwords"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lineId",
          "text",
          "bbox",
          "textSource",
          "uncertainWords"
        ],
        "title": "PageLine"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}